+++ src/rccl_wrap.cc
@@ -127,43 +127,14 @@ ncclResult_t rcclFuncMaxSendRecvCount(ncclFunc_t func, int nRanks, size_t count,
   return ncclSuccess;
 }
 
-//RCCL runtime param to set Unroll Factor
-RCCL_PARAM(UnrollFactor, "UNROLL_FACTOR", 0);
-
 ncclResult_t commSetUnrollFactor(struct ncclComm* comm) {
   hipDeviceProp_t devProp;
   CUDACHECK(hipGetDeviceProperties(&devProp, comm->cudaDev));
-
-  //If RCCL runtime param is set, it will override defaults, if supported
-  if (rcclParamUnrollFactor() != 0) {
-#if ENABLE_COLLTRACE
-    if(rcclGetKernelIndex(rcclParamUnrollFactor(), comm->collTraceEnabled)) {
-#else
-    if(rcclGetKernelIndex(rcclParamUnrollFactor(), false)) {
-#endif
-      comm->unroll = rcclParamUnrollFactor();
-      INFO(NCCL_INIT, "RCCL Unroll Factor (user-defined): %d", comm->unroll);
-      return ncclSuccess;
-    }
-    else {
-      // Fall back to default unroll
-      WARN("Requested RCCL_UNROLL_FACTOR: %ld is invalid and does not exist in `rcclKernelTable`. Falling back to pre-set unroll.", rcclParamUnrollFactor());
-    }
-  }
-
-  if (IsArchMatch(devProp.gcnArchName, "gfx950")) {
-    //on gfx950, use unroll=1 for single-node and unroll=2 for multi-node
-    if (comm->nNodes == 1)
-      comm->unroll = 1;
-    else
-      comm->unroll = 2;
-  }
-  else if((IsArchMatch(devProp.gcnArchName, "gfx908")) ||
-          (IsArchMatch(devProp.gcnArchName, "gfx942") && devProp.multiProcessorCount > 80))
-    //on MI300X and gfx908, use unroll=2
-    comm->unroll = 2;
+  if(IsArchMatch(devProp.gcnArchName, "gfx950"))
+    comm->unroll = NCCL_UNROLL_1;
+  else if(IsArchMatch(devProp.gcnArchName, "gfx908") || ((IsArchMatch(devProp.gcnArchName, "gfx942") && devProp.multiProcessorCount > 80)))
+    comm->unroll = NCCL_UNROLL_2;
   else
-    comm->unroll = 4;
-  INFO(NCCL_INIT, "RCCL Unroll Factor (pre-set): %d", comm->unroll);
+    comm->unroll = NCCL_UNROLL_4;
   return ncclSuccess;
 }